Adds metadata, adds commands
