+++ testing/tokenfactory/test_factory.sh
@@ -1,25 +1,31 @@
-# http://localhost:1317/clockworkgr/tokenfactory/tokenfactory/denom
+# TODO: prefix with owner name so no denoms can conflict with namespace? eveaddr/denom
+# TODO: dont allow minting token named eve / ueve?
 
-# TODO: denom & ticker = the same?
-# TODO:prefix with owner name so no denoms can conflict with namespace? eveaddr/denom
+eved tx tokenfactory create-denom aaa "my denom" 6 100000 false --from eve1 --chain-id eve-t1 --yes -b sync --keyring-backend test
+eved q tokenfactory show-denom aaa
+# eved q tokenfactory list-denoms
 
-eved tx tokenfactory create-denom reece "my denom" 6 100000 false --from eve1 --chain-id eve-t1 --yes -b sync --keyring-backend test
+# = Description =
+eved tx tokenfactory update-denom-desc aaa "new-desc-2 is here for you" --from eve1 --chain-id eve-t1 --yes -b sync --keyring-backend test
+# eved q tokenfactory show-denom aaa
 
-eved q tokenfactory list-denoms
+# = Token Image =
+eved tx tokenfactory update-denom-image aaa "https://raw.githubusercontent.com/cosmos/chain-registry/master/cosmoshub/images/atom.svg" --from eve1 --chain-id eve-t1 --yes -b sync --keyring-backend test
+# eved q tokenfactory show-denom aaa
 
-eved q tokenfactory show-denom reece
+# = Website =
+eved tx tokenfactory update-denom-website aaa "https://reece.sh" --from eve1 --chain-id eve-t1 --yes -b sync --keyring-backend test
+# eved q tokenfactory show-denom aaa
 
 
-# update denom - max-supply & can change should be optional
-# eved tx tokenfactory update-denom reece "new-desc-2" 100001 false --from eve1 --chain-id eve-t1 --yes -b sync --keyring-backend test
-eved tx tokenfactory update-denom-desc reece "new-desc-2" --from eve1 --chain-id eve-t1 --yes -b sync --keyring-backend test
 
-# add:
-# update-denom-url ?
-# update-denom-supply (increase/descrease, and toggle change to false) IF they are allowed too.
+# = Change Max Supply = # is this even needed?
+# 'Cannot revert change maxsupply' if false in create-denom
+eved tx tokenfactory update-denom-change-supply aaa true --from eve1 --chain-id eve-t1 --yes -b sync --keyring-backend test
+# eved q tokenfactory show-denom aaa
 
-# todo: dont allow minting token named eve / ueve?
 
-# mint tokens to eve1 account & check that they have the tokens now.
-eved tx tokenfactory mint-and-send-tokens reece 9 eve1hj5fveer5cjtn4wd6wstzugjfdxzl0xpysfwwn --from eve1 --chain-id eve-t1 --yes -b sync --keyring-backend test
-eved q bank balances eve1hj5fveer5cjtn4wd6wstzugjfdxzl0xpysfwwn --denom=reece
+# = Mint Tokens -> An Account =
+# key found in ./test_node.sh
+eved tx tokenfactory mint-and-send-tokens aaa 9 eve1hj5fveer5cjtn4wd6wstzugjfdxzl0xpysfwwn --from eve1 --chain-id eve-t1 --yes -b sync --keyring-backend test
+eved q bank balances eve1hj5fveer5cjtn4wd6wstzugjfdxzl0xpysfwwn --denom=aaa
