--- x/tokenfactory/client/cli/tx_update_owner.go
@@ -1,45 +1,0 @@
-package cli
-
-import (
-	"strconv"
-
-	"github.com/spf13/cobra"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/client/tx"
-	"github.com/notional-labs/eve/x/tokenfactory/types"
-)
-
-var _ = strconv.Itoa(0)
-
-func CmdUpdateOwner() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "update-owner [denom] [new-owner]",
-		Short: "Broadcast message UpdateOwner",
-		Args:  cobra.ExactArgs(2),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			argDenom := args[0]
-			argNewOwner := args[1]
-
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			msg := types.NewMsgUpdateOwner(
-				clientCtx.GetFromAddress().String(),
-				argDenom,
-				argNewOwner,
-			)
-			if err := msg.ValidateBasic(); err != nil {
-				return err
-			}
-			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
-		},
-	}
-
-	flags.AddTxFlagsToCmd(cmd)
-
-	return cmd
-}
